+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/components/chat_messages/ChatMessageListCellFactory.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.desktop.main.content.components.chatMessages;
+package bisq.desktop.main.content.components.chat_messages;
 
 import bisq.chat.ChatChannel;
 import bisq.chat.ChatMessage;
@@ -23,10 +23,10 @@ import bisq.desktop.main.content.bisq_easy.offerbook.MyOfferMessageBox;
 import bisq.desktop.main.content.bisq_easy.offerbook.PeerOfferMessageBox;
 import bisq.desktop.main.content.bisq_easy.open_trades.MyProtocolLogMessageBox;
 import bisq.desktop.main.content.bisq_easy.open_trades.PeerProtocolLogMessageBox;
-import bisq.desktop.main.content.components.chatMessages.messages.LeaveChatMessageBox;
-import bisq.desktop.main.content.components.chatMessages.messages.MessageBox;
-import bisq.desktop.main.content.components.chatMessages.messages.MyMessageBox;
-import bisq.desktop.main.content.components.chatMessages.messages.PeerMessageBox;
+import bisq.desktop.main.content.components.chat_messages.messages.LeaveChatMessageBox;
+import bisq.desktop.main.content.components.chat_messages.messages.MessageBox;
+import bisq.desktop.main.content.components.chat_messages.messages.MyMessageBox;
+import bisq.desktop.main.content.components.chat_messages.messages.PeerMessageBox;
 import javafx.geometry.Insets;
 import javafx.geometry.Pos;
 import javafx.scene.Node;
