+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/components/chat_messages/ChatMessagesComponent.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.desktop.main.content.components.chatMessages;
+package bisq.desktop.main.content.components.chat_messages;
 
 import bisq.bisq_easy.NavigationTarget;
 import bisq.chat.*;
