+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/components/chat_messages/messages/MessageBox.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.desktop.main.content.components.chatMessages.messages;
+package bisq.desktop.main.content.components.chat_messages.messages;
 
 import javafx.scene.layout.VBox;
